Implement presentation logic for fetching user reservations
Introduce `MyReservationViewModel` to handle the fetching, pagination, and state management of a user's confirmed reservations. Additionally, correct module namespaces and update the project's Kotlin version.

- **Presentation Layer:**
    - Create `MyReservationViewModel` to manage the reservation list state, including initial loading, pagination (`loadMore`), and error handling.
    - Add `ReservationTicketUI` data class and a mapping function (`toReservationTicketUI`) to convert domain `ReservationTicket` objects for UI consumption.
- **Domain & Data Layer:**
    - Change the `offset` parameter type in `MobileReservationRepository.getMyReservationTickets` (and its implementation) from `Long` to `Int`.
- **Build Configuration:**
    - Fix the Android namespaces in the `feature/reservation/data` and `feature/reservation/domain` `build.gradle.kts` files to correctly reflect their respective layers, replacing the incorrect `presentation` suffix.
    - Bump the Kotlin version to `2.4.10` in the version catalog.

+++ feature/reservation/data/src/commonMain/kotlin/com/martdev/flickq/feature/reservation/data/MobileReservationRepoImpl.kt
@@ -16,7 +16,7 @@ class MobileReservationRepoImpl(
     override suspend fun getMyReservationTickets(
         status: ReservationStatus,
         limit: Int,
-        offset: Long
+        offset: Int
     ): Result<List<ReservationTicket>, DataError> {
        return client.getData<List<ReservationTicketDTO>>(
             "/reservation/my-reservations",
